+++ backend/canisters/group/impl/src/updates/c2c_join_group.rs
@@ -1,5 +1,5 @@
 use crate::activity_notifications::handle_activity_notification;
-use crate::guards::caller_is_local_user_index;
+use crate::guards::caller_is_user_index_or_local_user_index;
 use crate::model::participants::AddResult;
 use crate::{mutate_state, read_state, run_regular_jobs, AddParticipantArgs, RuntimeState};
 use canister_api_macros::update_msgpack;
@@ -9,12 +9,12 @@ use gated_groups::{check_if_passes_gate, CheckIfPassesGateResult};
 use group_canister::c2c_join_group::{Response::*, *};
 use types::{CanisterId, EventIndex, GroupGate, MessageIndex, ParticipantJoined, UsersUnblocked};
 
-#[update_msgpack(guard = "caller_is_local_user_index")]
+#[update_msgpack(guard = "caller_is_user_index_or_local_user_index")]
 #[trace]
 async fn c2c_join_group(args: Args) -> Response {
     run_regular_jobs();
 
-    match read_state(|state| is_permitted_to_join(false, state)) {
+    match read_state(is_permitted_to_join) {
         Ok(Some((gate, user_index_canister_id))) => {
             match check_if_passes_gate(gate, args.user_id, user_index_canister_id).await {
                 CheckIfPassesGateResult::Success => {}
@@ -29,18 +29,18 @@ async fn c2c_join_group(args: Args) -> Response {
     mutate_state(|state| c2c_join_group_impl(args, state))
 }
 
-fn is_permitted_to_join(
-    has_passed_gate: bool,
-    runtime_state: &RuntimeState,
-) -> Result<Option<(GroupGate, CanisterId)>, Response> {
-    if runtime_state.data.is_frozen() {
+fn is_permitted_to_join(runtime_state: &RuntimeState) -> Result<Option<(GroupGate, CanisterId)>, Response> {
+    let caller = runtime_state.env.caller();
+
+    // If the call is from the user index then we skip the checks
+    if caller == runtime_state.data.user_index_canister_id {
+        Ok(None)
+    } else if runtime_state.data.is_frozen() {
         Err(ChatFrozen)
     } else if !runtime_state.data.is_accessible_by_non_member() {
         Err(GroupNotPublic)
     } else if let Some(limit) = runtime_state.data.participants.user_limit_reached() {
         Err(ParticipantLimitReached(limit))
-    } else if has_passed_gate {
-        Ok(None)
     } else {
         Ok(runtime_state
             .data
@@ -51,13 +51,6 @@ fn is_permitted_to_join(
 }
 
 fn c2c_join_group_impl(args: Args, runtime_state: &mut RuntimeState) -> Response {
-    if runtime_state.data.gate.is_some() {
-        // We need to check again because things may have changed during the async gate check
-        if let Err(response) = is_permitted_to_join(true, runtime_state) {
-            return response;
-        }
-    }
-
     let now = runtime_state.env.now();
     let min_visible_event_index;
     let min_visible_message_index;