+++ backend/canisters/user_index/api/src/queries/mod.rs
@@ -2,6 +2,7 @@ pub mod c2c_lookup_user;
 pub mod check_username;
 pub mod current_user;
 pub mod platform_moderators;
+pub mod platform_moderators_group;
 pub mod platform_operators;
 pub mod referral_leaderboard;
 pub mod referral_metrics;