+++ backend/canisters/user_index/api/src/updates/mod.rs
@@ -2,6 +2,7 @@ pub mod add_local_user_index_canister;
 pub mod add_platform_moderator;
 pub mod add_platform_operator;
 pub mod add_referral_codes;
+pub mod assign_platform_moderators_group;
 pub mod c2c_migrate_user_principal;
 pub mod c2c_notify_events;
 pub mod c2c_notify_low_balance;