+++ backend/canisters/user_index/impl/src/queries/mod.rs
@@ -3,6 +3,7 @@ pub mod check_username;
 pub mod current_user;
 pub mod http_request;
 pub mod platform_moderators;
+pub mod platform_moderators_group;
 pub mod platform_operators;
 pub mod referral_leaderboard;
 pub mod referral_metrics;