+++ backend/canisters/user_index/impl/src/queries/platform_moderators_group.rs
@@ -1,0 +1,12 @@
+use crate::{read_state, RuntimeState};
+use ic_cdk_macros::query;
+use user_index_canister::platform_moderators_group::{Response::*, *};
+
+#[query]
+fn platform_moderators_group(_args: Args) -> Response {
+    read_state(platform_moderators_group_impl)
+}
+
+fn platform_moderators_group_impl(runtime_state: &RuntimeState) -> Response {
+    Success(runtime_state.data.platform_moderators_group.unwrap())
+}