+++ frontend/openchat-agent/src/services/localUserIndex/candid/idl.js
@@ -327,6 +327,7 @@ export const idlFactory = ({ IDL }) => {
   const NnsPendingCryptoTransaction = IDL.Record({
     'to' : NnsUserOrAccount,
     'fee' : IDL.Opt(Tokens),
+    'created' : TimestampNanos,
     'token' : Cryptocurrency,
     'memo' : IDL.Opt(Memo),
     'amount' : Tokens,
@@ -334,6 +335,7 @@ export const idlFactory = ({ IDL }) => {
   const SnsPendingCryptoTransaction = IDL.Record({
     'to' : Icrc1Account,
     'fee' : Tokens,
+    'created' : TimestampNanos,
     'token' : Cryptocurrency,
     'memo' : IDL.Opt(Memo),
     'amount' : Tokens,