+++ frontend/openchat-agent/src/services/online/candid/types.d.ts
@@ -595,6 +595,7 @@ export type NnsNeuronId = bigint;
 export interface NnsPendingCryptoTransaction {
   'to' : NnsUserOrAccount,
   'fee' : [] | [Tokens],
+  'created' : TimestampNanos,
   'token' : Cryptocurrency,
   'memo' : [] | [Memo],
   'amount' : Tokens,
@@ -802,6 +803,7 @@ export type SnsNeuronId = Uint8Array | number[];
 export interface SnsPendingCryptoTransaction {
   'to' : Icrc1Account,
   'fee' : Tokens,
+  'created' : TimestampNanos,
   'token' : Cryptocurrency,
   'memo' : [] | [Memo],
   'amount' : Tokens,