+++ frontend/openchat-agent/src/services/user/candid/idl.js
@@ -365,6 +365,7 @@ export const idlFactory = ({ IDL }) => {
   const NnsPendingCryptoTransaction = IDL.Record({
     'to' : NnsUserOrAccount,
     'fee' : IDL.Opt(Tokens),
+    'created' : TimestampNanos,
     'token' : Cryptocurrency,
     'memo' : IDL.Opt(Memo),
     'amount' : Tokens,
@@ -372,6 +373,7 @@ export const idlFactory = ({ IDL }) => {
   const SnsPendingCryptoTransaction = IDL.Record({
     'to' : Icrc1Account,
     'fee' : Tokens,
+    'created' : TimestampNanos,
     'token' : Cryptocurrency,
     'memo' : IDL.Opt(Memo),
     'amount' : Tokens,