+++ frontend/openchat-agent/src/services/userIndex/mappers.ts
@@ -124,6 +124,9 @@ export function registerUserResponse(candid: ApiRegisterUserResponse): RegisterU
     if ("PublicKeyInvalid" in candid) {
         return "public_key_invalid";
     }
+    if ("ReferralCodeInvalid" in candid) {
+        return "referral_code_invalid";
+    }
 
     throw new UnsupportedValueError("Unexpected ApiRegisterUserResponse type received", candid);
 }