+++ frontend/openchat-client/src/openchat.ts
@@ -338,7 +338,7 @@ export class OpenChat extends EventTarget {
     private _lastOnlineDatesPromise: Promise<Record<string, number>> | undefined;
     private _cachePrimer: CachePrimer | undefined = undefined;
     private _membershipCheck: number | undefined;
-    private _referredBy: string | undefined = undefined;
+    private _referralCode: string | undefined = undefined;
 
     constructor(private config: OpenChatConfig) {
         super();
@@ -1144,7 +1144,8 @@ export class OpenChat extends EventTarget {
     deleteMessage(
         chatId: string,
         threadRootMessageIndex: number | undefined,
-        messageId: bigint
+        messageId: bigint,
+        asPlatformModerator?: boolean
     ): Promise<boolean> {
         const chat = this._liveState.chatSummaries[chatId];
 
@@ -1182,7 +1183,7 @@ export class OpenChat extends EventTarget {
         }
 
         return this.api
-            .deleteMessage(chatType, chatId, messageId, threadRootMessageIndex)
+            .deleteMessage(chatType, chatId, messageId, threadRootMessageIndex, asPlatformModerator)
             .then((resp) => {
                 const success = resp === "success";
                 if (!success) {
@@ -2479,7 +2480,7 @@ export class OpenChat extends EventTarget {
     }
 
     private canRetryMessage(content: MessageContent): boolean {
-        return content.kind !== "crypto_content" && content.kind !== "poll_content";
+        return content.kind !== "poll_content";
     }
 
     sendMessageWithAttachment(
@@ -2904,12 +2905,12 @@ export class OpenChat extends EventTarget {
             localStorage.setItem("openchat_referredby", code);
             captured = true;
         }
-        this._referredBy = localStorage.getItem("openchat_referredby") ?? undefined;
+        this._referralCode = localStorage.getItem("openchat_referredby") ?? undefined;
         return captured;
     }
 
     registerUser(username: string): Promise<RegisterUserResponse> {
-        return this.api.registerUser(username, this._referredBy).then((res) => {
+        return this.api.registerUser(username, this._referralCode).then((res) => {
             if (res === "success") {
                 localStorage.removeItem("openchat_referredby");
             }