+++ frontend/openchat-shared/src/domain/chat/chat.ts
@@ -50,6 +50,7 @@ export type PendingCryptocurrencyWithdrawal = {
     amountE8s: bigint;
     feeE8s?: bigint;
     memo?: bigint;
+    createdAtNanos: bigint;
 };
 
 export type CompletedCryptocurrencyWithdrawal = {
@@ -102,6 +103,7 @@ export type PendingCryptocurrencyTransfer = {
     amountE8s: bigint;
     feeE8s?: bigint;
     memo?: bigint;
+    createdAtNanos: bigint;
 };
 
 export type FailedCryptocurrencyTransfer = {
@@ -1532,7 +1534,9 @@ export type DeleteMessageResponse =
     | "success"
     | "message_not_found"
     | "user_suspended"
-    | "chat_frozen";
+    | "chat_frozen"
+    | "internal_error"
+    | "not_platform_moderator";
 
 export type UndeleteMessageResponse =
     | {