+++ frontend/openchat-shared/src/domain/worker.ts
@@ -369,7 +369,7 @@ type SubscriptionExists = Request<{
 
 type RegisterUser = Request<{
     username: string;
-    referredBy: string | undefined;
+    referralCode: string | undefined;
 }> & {
     kind: "registerUser";
 };
@@ -472,6 +472,7 @@ type DeleteMessage = Request<{
     chatId: string;
     messageId: bigint;
     threadRootMessageIndex?: number;
+    asPlatformModerator?: boolean;
 }> & {
     kind: "deleteMessage";
 };