+++ backend/canisters/group/CHANGELOG.md
@@ -8,6 +8,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 
 - Use hardcoded ledger ids ([#3452](https://github.com/open-ic/open-chat/pull/3452))
+- Allow platform moderators to delete any messages ([#3491](https://github.com/open-ic/open-chat/pull/3491))
+- Allow the user index to add group members ([#3493](https://github.com/open-ic/open-chat/pull/3493))
+- Added `created` to pending transactions ([#3494](https://github.com/open-ic/open-chat/pull/3494))
 
 ## [[2.0.662](https://github.com/open-ic/open-chat/releases/tag/v2.0.662-group)] - 2023-04-16